Move birnn implementation to birnn file and create utils file
