+++ distilled/main.cpp
@@ -2,23 +2,8 @@
 #include <vector>
 
 #include "CLI/CLI.hpp"
-#include "layers/generic.h"
-#include "marian.h"
-#include "models/decoder.h"
-#include "models/encoder.h"
-#include "models/model_factory.h"
-#include "models/s2s.h"
 
-using namespace marian;
-using Exprs = std::vector<std::pair<std::string, Expr> >;
-
-void createGraphParams(Ptr<ExpressionGraph>& graph, const int dim_vocab, const int dim_emb) {
-  graph->param("embeddings_txt_src", {dim_vocab, dim_emb}, inits::glorotUniform());
-}
-
-void saveResults(const std::string& filePath, const Exprs& exprs);
-Exprs forward(Ptr<ExpressionGraph>& graph, const std::vector<WordIndex>& tokens_src, const int dim_emb,
-              const std::vector<float>& mask_src);
+#include "birnn.h"
 
 int main(const int argc, const char* argv[]) {
   std::string modelPath;
@@ -54,9 +39,6 @@ int main(const int argc, const char* argv[]) {
   auto tokens_src = toWordIndexVector(tokens_word_src);
   std::vector<float> mask_src(tokens_src.size(), 1.0);
 
-  // // Create randon params
-  // forward(graph, tokens_src, dim_emb, mask);
-
   // Load converted python model
   graph->load(modelPath);
 
@@ -68,65 +50,11 @@ int main(const int argc, const char* argv[]) {
   mask_src = {1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0,
               1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0};
 
-  // createGraphParams( graph, dim_vocab, dim_emb);
-
-  const auto results = forward(graph, tokens_src, dim_emb, mask_src);
+  const auto results = distilled::birnn::forward(graph, tokens_src, dim_emb, mask_src);
 
   if (!outputPath.empty()) {
-    saveResults(outputPath, results);
+    distilled::saveResults(outputPath, results);
   }
 
   return 0;
-}
-
-Exprs forward(Ptr<ExpressionGraph>& graph, const std::vector<WordIndex>& tokens_src, const int dim_emb,
-              const std::vector<float>& mask_src) {
-  const int dim_batch = 1;
-  const int dim_tokens_src = tokens_src.size();
-
-  // Source Embedding
-  auto embeddings_txt_src = graph->get("embeddings_txt_src");
-
-  auto embedded_text_src = reshape(rows(embeddings_txt_src, tokens_src), {dim_batch, dim_tokens_src, dim_emb});
-
-  // Source Seq2seq Encoder
-  auto options =
-      New<Options>("enc-depth", 1, "dropout-rnn", 0.0f, "enc-cell", "gru", "dim-rnn", dim_emb, "layer-normalization",
-                   false, "skip", false, "enc-cell-depth", 1, "prefix", "encoder_s2s_text_src", "hidden-bias", true);
-
-  marian::EncoderS2S encoderS2S(graph, options);
-
-  auto embedded_text_src_t = transpose(embedded_text_src, {1, 0, 2});
-
-  auto mask_src_expr = graph->constant({static_cast<int>(mask_src.size()), 1, 1}, marian::inits::fromVector(mask_src));
-
-  auto encoded_text_src_t = encoderS2S.applyEncoderRNN(graph, embedded_text_src_t, mask_src_expr, "bidirectional");
-
-  auto encoded_text_src = transpose(encoded_text_src_t, {1, 0, 2});
-
-  // Source Linear Layer
-  auto linear_layer_src_weight = graph->get("linear_layer_src_weight");
-  auto linear_layer_src_bias = graph->get("linear_layer_src_bias");
-
-  auto encoded_text_src_linear_op = dot(encoded_text_src, linear_layer_src_weight, false, true) + linear_layer_src_bias;
-
-  std::cout << graph->graphviz() << std::endl;
-
-  graph->forward();
-
-  return {{"embedded_text_src", embedded_text_src},
-          {"encoded_text_src", encoded_text_src},
-          {"encoded_text_src_linear_op", encoded_text_src_linear_op}};
-}
-
-void saveResults(const std::string& filePath, const Exprs& exprs) {
-  std::vector<io::Item> items;
-
-  std::transform(std::begin(exprs), std::end(exprs), std::back_inserter(items), [](const auto& expr) {
-    io::Item item;
-    expr.second->val()->get(item, expr.first);
-    return item;
-  });
-
-  io::saveItems(filePath, items);
 }
